requirements.txt
streamlit>=1.35.0
PyYAML>=6.0.1
pytest>=8.2.0
